100-days-of-algorithms/src/main/java/com/algos/array/NumbersDisapperaredInArray.java
package com.algos.array;

import java.util.ArrayList;
import java.util.List;

public class NumbersDisapperaredInArray {

    public List<Integer> findDisappearedNumbers(int[] nums) {
        List<Integer> result = new ArrayList<>();
        if(nums == null || nums.length <= 1) {
            return result;
        }

        for(int i = 0; i < nums.length; i++) {
            int index = Math.abs(nums[i]) - 1;

            if(nums[index] > 0) {
                nums[index] = -nums[index];
            }
        }

        for(int i = 0; i < nums.length; i++) {
            if(nums[i] > 0) {
                result.add(i + 1);
            }
        }

        return result;
    }
}
